move more period dtype tests
